EdgeQA/requirements.txt
playwright>=1.44.0
pytest>=8.0.0
pytest-xdist>=3.6.0
allure-pytest>=2.13.5
pytest-html>=4.1.1
pyyaml>=6.0.1
openpyxl>=3.1.2
requests>=2.32.0
jsonschema>=4.22.0
